feat: remove base styles, style per-element
In a lot of places I ended up undoing these base styles for
<h1>s, <p>s, <a>s, etc. Using flat rules across the board
like this is discouraged by Tailwind, and I'm understaning
why. More often than not, a page needs its own rules for its
elements.

Flat styles also make it harder to write semantic HTML --
we shouldn't be using e.g. h1, h2, or h3 as the way to select
a specific style of heading.