feat: add handshake lab page
The handshake package comes with a set of images that are used as
textures for the arm and hand objects. Matter-js expects these images to
be fetch-able from the web root.

The simplest way I could think of doing this is just copying the images
from node_modules to SvelteKit's static folder, in a postinstall script.
I'm not sure if there's a better way of doing this?

I've also added a couple of global styles ot app.css that I'm pretty sure
I'll want everywhere.